Build `mozc_tip64.dll` for ARM64 with GYP
As a preparation to build Mozc for ARM64 in Windows (#1130), this commit
demonstrates that it is technically possible for us to build
'mozc_tip64.dll' as an ARM64 DLL with GYP even though GYP itself does
not officially support it.

With this commit you need to install the following additional Visual
Studio components even when you do not build Mozc for ARM64 Windows.

  * Microsoft.VisualStudio.Component.VC.Redist.14.Latest
  * Microsoft.VisualStudio.Component.VC.ATL
  * Microsoft.VisualStudio.Component.VC.Tools.x86.x64
  * Microsoft.VisualStudio.Component.VC.ATL.ARM64
  * Microsoft.VisualStudio.Component.VC.Tools.ARM64

Note that 'mozc_tip64.dll' for ARM64 will be built only when explicitly
specified in 'build_mozc.py build'. There must be no change in the final
artifacts when just building the installer package.

Note also that we still need to build ARM64X DLL to actually support
ARM64 Windows environment.