Allow to build ARM64 native installer
When built on ARM64 Windows, the installer will be built as an ARM64
binary unless '--platforms=//:windows_x86_64' is explicitly specified.

When built on x64 Windows, the installer will be built as an x64 binary
unless '--platforms=//:windows_arm64' is explicitly specified.

When built for ARM64, the following files will be built as ARM64
binaries:

 * mozc_broker.exe
 * mozc_cache_service.exe
 * mozc_installer_helper.dll
 * mozc_renderer.exe
 * mozc_server.exe
 * mozc_tool.exe
 * Qt6Core.dll
 * Qt6Gui.dll
 * Qt6Widgets.dll
 * qwindows.dll

Note that TIP for ARM64 is not yet included in the installer.