Actually require Windows 10+ in the installer
While Windows 10 has been our official minimum supported version for a
while, the installer did not enforce this requirement strictly because
'WindowsNT' variable, which is available in WiX script by default, does
not work to distinguish Windows versions since Windows 8.1.

To address this limitation, we start using the OS build number, which
looks to be sufficient to compare API revisions across Windows 10/11
releases and Windows Server products. Here we use registry to obtain
the build number. This is supposed to be safe as even Microsoft products
(Intune [1] and FSLogix [2]) mentions the registry value. There are also
other code examples in Microsoft PowerToys [3] and Google bindiff [4].

Just to be extra safe, the condition we use ensures that the user can
still uninstall Mozc after upgrading Windows to a future version where
the registry value we use is somehow no longer available.

 [1]: https://learn.microsoft.com/en-us/intune/intune-service/apps/apps-win32-deploy-update-package
 [2]: https://learn.microsoft.com/en-us/fslogix/reference-configuration-settings#custom-environment-variables
 [3]: https://github.com/microsoft/PowerToys/blob/ca473b488bc3df626140cd332b9a1ea4ff1fcdf2/installer/PowerToysSetup/Product.wxs#L43-L45
 [4]: https://github.com/google/bindiff/blob/7e4d0375035b2353f22f44311f747eedafd98c4e/packaging/msi/Setup.wxs#L70-L75