Also update `last_down_key` in `KeyEventHandler::ImeToAsciiEx`
As explained in #1415 and code comment copied in this commit, Mozc
clients are expected to send isolated modifier key press events to
mozc_server. This has actually been working fine in IMM32 version of
Mozc client, but it turns out that the TSF version of Mozc client may
not work as expected. To understand the root cause of the problem, we
need to look into how key events are processed in IMM32 and TSF.

  |     IMM32     |    Text Services Framework (TSF)    |
  |---------------|-------------------------------------|
  | ImeProcessKey | ITfKeyEventSink::OnTestKey{Down,Up} |
  | ImeToAsciiEx  | ITfKeyEventSink::OnKey{Down,Up}     |

Mozc's KeyEventHandler still uses the legacy IMM32 naming as it had
been shared between IMM32 and TSF clients until we dropped support for
IMM32 client.

The problem is that our KeyEventHandler::ImeProcessKey is written with
an assumption that it is guaranteed to be called before
KeyEventHandler::ImeToAsciiEx, but in TSF, this assumption is no longer
valid. In TSF, there are three possible scenarios for a key down event:

 A. ITfKeyEventSink::OnTestKeyDown gets called but not consumed by TIP.
    In this case, ITfKeyEventSink::OnKeyDown will not be called thus
    OnTestKeyDown is the last chance to memorize the key down event.
 B. ITfKeyEventSink::OnTestKeyDown gets called and consumed by TIP. In
    this case, ITfKeyEventSink::OnKeyDown will be called for the same
    key event.
 C. ITfKeyEventSink::OnTestKeyDown is skipped. In this case,
    ITfKeyEventSink::OnKeyDown will be the only chance to memorize the
    key down event.

We need to handle all the three scenarios correctly to ensure that
isolated modifier key press events are sent to mozc_server as expected.

In order to minimize the risk of regression, this commit updates
last_down_key in both KeyEventHandler::ImeProcessKey and
KeyEventHandler::ImeToAsciiEx.

Maybe it will be possible to make KeyEventHandler::ImeProcessKey no-op
and always return "consumed" for simplicity in the future.

Closes #1415.